JFileChooser and Excel Graphs
Created a JFileChooser object to save the Excel graphs. Need to implement a loop to ensure that the user can select and deselect a file as many times as they want.

Created two more Excel graphs for statistical data: Max Reps and Max Volume. Need to attempt to make a loop to create all 3 graphs to cut down on duplicate code.

+++ FitnessAppJavaFX/src/model/WorkoutTracker.java
@@ -168,7 +168,7 @@ public class WorkoutTracker {
 				int weight = current.indexOf("Weight:");
 
 				this.workoutList.get(workoutNum).getExercise(exerciseNum)
-						.addSet(new Set(Integer.parseInt(current.substring(reps + 6, weight - 1)),
+						.addSet(new Set(Double.parseDouble(current.substring(reps + 6, weight - 1)),
 								Double.parseDouble(current.substring(weight + 7, current.length()))));
 
 			}
